+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## [0.8.0] - dev
+
+### Added
+
+- MySql structure.
+- Data Types for DBs.
+- Mongo#Types property.
+
+### Changed
+
+- DB structure renamed to Mongo.
+- Client variable should be defined in PermissionLevels#test method if it is needed.
+- You can use client argument in PermissionLevels#add method.
+
+### Fixed
+
 ## [0.7.0] - 2019-08-16
 
 ### Added